We no longer need to filter for Herts libraries
Before: the `herts.spydus.co.uk` instance would show availability info
from branches in multiple counties, including Herts and Bucks.  To tell
them apart, each branch name would be shown with a suffix in parens, e.g.

    St Albans Library (Hertfordshire County Council)

The old code would filter to look for branches with the Herts suffix.

One of the changes in the Spydus upgrade is that `herts.spydus.co.uk` now
only shows books from the Hertfordshire branches, which makes this filtering
unnecessary.  But the filter was still being applied, so it would ignore
everything.